refactor: move splash removal to core provider
Extract splash screen removal logic from App component into a
provideSplashRemoval() function in core/ to keep App as a minimal shell
and respect architecture boundary rules. Includes comprehensive tests.

- New: core/splash-removal.ts with provideSplashRemoval() provider
- New: core/splash-removal.spec.ts with 6 unit tests
- Revert: App.ts to minimal shell component
- Delete: app.spec.ts (tests moved to core)
- Revert: ESLint boundary rule (no longer needed)
- Fix: Add error callback to RxJS subscribe for robustness
