+++ frontend/projects/webapp/src/app/core/splash-removal.spec.ts
@@ -1,6 +1,9 @@
 import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest';
 import { TestBed } from '@angular/core/testing';
-import { provideZonelessChangeDetection } from '@angular/core';
+import {
+  ApplicationInitStatus,
+  provideZonelessChangeDetection,
+} from '@angular/core';
 import {
   Router,
   NavigationEnd,
@@ -10,42 +13,42 @@ import {
   provideRouter,
 } from '@angular/router';
 import { Subject } from 'rxjs';
-import { App } from './app';
+import { provideSplashRemoval } from './splash-removal';
 
-describe('App', () => {
+describe('provideSplashRemoval', () => {
   let routerEvents$: Subject<unknown>;
   let splashElement: HTMLDivElement;
 
   beforeEach(() => {
     routerEvents$ = new Subject();
 
-    // Create a fake splash element in the DOM
     splashElement = document.createElement('div');
     splashElement.id = 'pulpe-splash';
     document.body.appendChild(splashElement);
+  });
 
+  afterEach(() => {
+    document.getElementById('pulpe-splash')?.remove();
+  });
+
+  function setup(): Promise<void> {
     TestBed.configureTestingModule({
-      imports: [App],
       providers: [
         provideZonelessChangeDetection(),
         provideRouter([]),
         {
           provide: Router,
-          useValue: {
-            events: routerEvents$.asObservable(),
-          },
+          useValue: { events: routerEvents$.asObservable() },
         },
+        provideSplashRemoval(),
       ],
     });
-  });
 
-  afterEach(() => {
-    // Clean up splash element if still present
-    document.getElementById('pulpe-splash')?.remove();
-  });
+    return TestBed.inject(ApplicationInitStatus).donePromise;
+  }
 
   it('should remove splash on NavigationEnd', async () => {
-    TestBed.createComponent(App);
+    await setup();
 
     routerEvents$.next(new NavigationEnd(1, '/', '/'));
     await waitForAnimationFrame();
@@ -54,7 +57,7 @@ describe('App', () => {
   });
 
   it('should remove splash on NavigationError', async () => {
-    TestBed.createComponent(App);
+    await setup();
 
     routerEvents$.next(new NavigationError(1, '/', new Error('fail')));
     await waitForAnimationFrame();
@@ -63,7 +66,7 @@ describe('App', () => {
   });
 
   it('should remove splash on NavigationCancel', async () => {
-    TestBed.createComponent(App);
+    await setup();
 
     routerEvents$.next(new NavigationCancel(1, '/', ''));
     await waitForAnimationFrame();
@@ -72,7 +75,7 @@ describe('App', () => {
   });
 
   it('should NOT remove splash on NavigationStart', async () => {
-    TestBed.createComponent(App);
+    await setup();
 
     routerEvents$.next(new NavigationStart(1, '/'));
     await waitForAnimationFrame();
@@ -83,11 +86,9 @@ describe('App', () => {
   it('should remove splash on timeout if no navigation event fires', async () => {
     vi.useFakeTimers();
 
-    TestBed.createComponent(App);
+    await setup();
 
-    // Advance past the 15s timeout
     vi.advanceTimersByTime(15_000);
-    // Flush the requestAnimationFrame queued by removeSplash
     vi.advanceTimersByTime(16);
 
     vi.useRealTimers();
@@ -98,12 +99,11 @@ describe('App', () => {
   it('should only remove splash once even if multiple events fire', async () => {
     const removeSpy = vi.spyOn(splashElement, 'remove');
 
-    TestBed.createComponent(App);
+    await setup();
 
     routerEvents$.next(new NavigationEnd(1, '/', '/'));
     await waitForAnimationFrame();
 
-    // Second event should not cause issues (element already removed)
     routerEvents$.next(new NavigationEnd(2, '/other', '/other'));
     await waitForAnimationFrame();
